requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.2
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
sqlparse==0.4.2
tomli==2.0.1
